update: user role in pulumi and added as default
